begin place rewrite for multiplace.
multiplace places within short range of a player for high damage in a short period of time. this is at the expense of using more crystals than usual. instead of taking into account the damage it might do to the entity as the deciding factor, it looks at how close the crystal is. it also automatically places crystals if the enemy's blockpos distance to a crystal is less than 3 blocks away and the y is the same. for optimum popping.