Fix overfilling inventory when using AutoObsidian
Check the maximum possible ender chests we can break, then pick the
minimum from that and the target stacks. Also continually check if our
inventory is full, and if it is transition to the done state (this is
mainly only useful if we somehow pick up other items while breaking echests).