Fixed multiple module descriptions (#355)
* Improved descriptions 

Made descriptions clearer and more concise.

* Improved description

Changed description to describe the function of the module and removed punctuation to fit the overall description style.

* Improved description

Changed a repeat use of the word in to within, to stay grammatically correct, and removed punctuation to stay within the style of other module's descriptions.

* Fixed description

The description seemed to be mistakenly copied over from logout logger, which was a cause of confusion. Removed the incorrect description and wrote a new correct one.

* Improved description

Changed might to attempt, as might is more informal and does not fit the use case as well.

* Improved description

Edited description as the module only allows you to add delay, not cancel.

* Improved description

Changed to be more formal and shorter

* Improved description

Changed description to describe the function of the module and be more explanatory.

* Updated description

Improved description and made it more descriptive.