Updated the Github issue templates (#396)
* Updated the incompatibilty report form

Updated it by adding more detailed questions, making it have better wording, and rewrote some sentences. This will make submitting a Incompatibility report easier, and assessing the report easier as well.

* Updated the feature request template 

Improved the wording and grammar and overall made it easier for someone to fill out, and after the fact have someone analyze and accurately determine the request.

* Updated the enchancment template

It seems like this was copied right from the feature request template, so I improved both wording and grammar, and also edited the questions slightly to be more enchantment specific.

* Enhanced the bug report template

Just improved wording a little bit.