Nametags: Add TOTEM_POP_COUNT ContentType
This patch allows for the popped totems to be able to be shown in the
nametags. The totem pop counting functionality is embedded in the
TotemPopCounter module, hence, it needs to be active in order for this
to work.

As to ensure this is the case, this patch adds some logic preventing
this feature to be enabled when the TotemPopCounter module is inactive,
and also re-enables the TotemPopCounter module if this feature is
enabled and the user tries to disable the TotemPopCounter module.