StashLogger: Add logToFile option
This patch adds the logToFile option in the StashLogger module. This
option, which is enabled by default, appends all the newfound stashes to
a JSON file ($lambda_folder/stash-logger/servername.json).

This is useful for when you are afk stash hunting and you would like to
keep a log of all the found stashes :)