Updated a bundle of module descriptions (#473)
* Updated BedAura module description

Just updated the wording a little

* Updated MidClickPearl module description

Shortened the description a little bit, removed automatically entirely from the sentence as I believe it is self explanatory without it.

* Updated module description

Just changed Modify to Modifies as that is a better word to use in this scenario, being the correct word used to refer to another person, or in this scenario a program/module!

* Updated module description

Extremely minor, just swapped when to from. Just a better word overall to use in this context

* Updated module description

Auto remount allows you to automatically remount not just a horse, but a boat, donkey, pig, and a few more entities, so I changed horse to rideable entities to better describe the function of the module.