[1.20.4] Feat: New gui, texture api and buffer api (#98)
Related: #79 #78

This pull request is the result of 7 months of hard work

A new maintainable click gui api was implemented
A new setting layout builder was added to create custom gui elements
such as buttons, sliders, etc.

The opengl common buffer was reworked to reduce overhead

The texture uploading process was simplified for the end user by
providing a new `TextureOwner` object that keeps track of all the
textures created in a given scope
The functions to upload textures are as follow:
- `fun Any.upload() = ...`
- `fun Any.uploadGif() = ...`
- `fun Any.uploadField() = ...`

A new AnimatedTexture class was added for gif rendering
The texture base class was extended to correctly handle other pixel
formats such as BGRA, RG, RED and provide mappings to mojang's native
image formats
The pixel buffer was reworked according to the buffer api

Both fonts, text and emojis, were reworked according to the texture api
A new [SDF
texture](https://steamcdn-a.akamaihd.net/apps/valve/2007/SIGGRAPH2007_AlphaTestedMagnification.pdf)
(also known as alpha-tested magnification) was implemented in order to
display clearer characters at any scale or transformation

There are many breaking changes that need to be corrected before merging
This pull request is NOT ready to be merged

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
  - Added animated texture support for GIFs with a new HUD module.
  - Introduced a Map Downloader that saves in‑game maps as PNG images.

- **UI Enhancements**
- Revamped Click GUI with updated layout, new settings (title bar
height, module size, round radius, etc.) and improved visual
customization.
- Enhanced font and emoji rendering using new shader techniques for
smoother, sharper text.

- **Improvements**
  - Streamlined module list and HUD rendering for a cleaner display.
  - Improved interactive animations and slider precision.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Blade-gl <bladecore.kt@gmail.com>
Co-authored-by: Constructor <fractalminds@protonmail.com>
Co-authored-by: blade <blade@gmail.com>
Co-authored-by: blade.kt <kurtasanoff@gmail.com>