Fix: Break Manager (#138)
This introduces a more immutable style system for item swapping logic.
Each tick a new SwapInfo is created for all active break infos.

This also fixes an issue with redundancy spam due to the build tasks
request logic.

Along with the fixes, ive added a new "Server Swap" setting for the user
to edit how many ticks the server should be given to recognise the
selected tool. This is hard limited to a minimum of 2 ticks for
secondary breaks to allow the user to select 0 for servers < 1.21.5 that
dont apply certain enchantments, like efficiency, as player attributes